Update index.html
The Leaflet Time-Slider plugin is meant to control the display of markers, one at a time, via sliding control. To accomplish this requires adding additional JS scripts and a stylesheet to your <head> section. The first set of these, the jquery UI script and stylesheet, is available via jquery's CDN: